moved game.js into src dir and built new version also moved css into style.css
