+++ css/styles.css
@@ -1,4 +1,157 @@
 /** CSS styles */
-.game-title {
-  color: red;
+div.scrolling-limit
+{
+  width: 1400px;
+  height: 30px;
+  background-color: red;
+    overflow:hidden;
+  white-space:nowrap;
+   margin-top: -35px;
+    margin-bottom: 0px;
+    margin-left: -10px;
+    margin-right: 0px;
+}
+
+body {
+  background: url('../img/3.jpg');
+  background-size: 1000px;
+}
+
+#points {
+  text-align: center;
+  font: 300% 'Trebuchet MS', cursive;
+  color: #fff;
+  text-shadow: 
+    0 0 10px #fff, 
+    0 0 30px #fff, 
+    0 0 50px #e3e98e, 
+    0 0 70px #e3e98e,
+    0 0 90px #e3e98e;
+}
+
+#purchaseTree1
+{  
+ display: block;
+    margin-top: -230px;
+    margin-bottom: 0px;
+    margin-left: 300px;
+    margin-right: 0px;
+  width: 150px;
+  height: 50px;
+    cursor: pointer;
+}
+
+#purchaseCar1
+{  
+ display: block;
+    margin-top: 25px;
+    margin-bottom: 0px;
+    margin-left: 300px;
+    margin-right: 0px;
+  width: 150px;
+  height: 50px;
+    cursor: pointer;
+}
+
+#purchaseSolar1
+{  
+ display: block;
+    margin-top: 25px;
+    margin-bottom: 0px;
+    margin-left: 300px;
+    margin-right: 0px;
+  width: 150px;
+  height: 50px;
+    cursor: pointer;
+}
+
+#purchaseFarm1
+{  
+ display: block;
+    margin-top: -200px;
+    margin-bottom: 0px;
+    margin-left: 900px;
+    margin-right: 0px;
+  width: 150px;
+  height: 50px;
+    cursor: pointer;
+}
+
+#purchaseGas1
+{  
+ display: block;
+    margin-top: 25px;
+    margin-bottom: 0px;
+    margin-left: 900px;
+    margin-right: 0px;
+  width: 150px;
+  height: 50px;
+    cursor: pointer;
+}
+
+#purchaseCapture1
+{  
+ display: block;
+    margin-top: 25px;
+    margin-bottom: 0px;
+    margin-left: 900px;
+    margin-right: 0px;
+  width: 150px;
+  height: 50px;
+    cursor: pointer;
+}
+
+#pollution
+{  
+    margin-top: -320px;
+    margin-bottom: 0px;
+    margin-left: 480px;
+    margin-right: 0px;
+  pointer-events:none;
+}
+
+#clicker {
+  width: 250px;
+  height: 250px;
+  background: url('../img/2.jpg');
+  border-radius: 50%;
+  background-size: 500px;
+  box-shadow: inset 16px 0 40px 6px rgb(0, 0, 0),
+    inset -3px 0 6px 2px rgba(255, 255, 255, 0.2);
+  animation-name: rotate;
+  animation-duration: 20s;
+  animation-iteration-count: infinite;
+  animation-timing-function: linear;
+   margin-top: 40px;
+    margin-bottom: 0px;
+    margin-left: 553px;
+    margin-right: 0px;
+}
+
+@keyframes rotate {
+  from { background-position-x: 0px; }
+  to { background-position-x: 210px; }
+}
+
+#newsfeed {
+  width: 1400px;
+  height: 30px;
+  color: white;
+  text-align: center;
+  font-size: 20px;
+  margin-top: 60px;
+  margin-bottom: 0px;
+  margin-left: 300px;
+    margin-right: 0px;
+  
+   position: relative;
+  -webkit-animation: scroll 20s infinite linear;;
+  -moz-animation: scroll 20s infinite linear;
+  -o-animation: scroll 20s infinite linear;
+  animation: scroll 20s infinite linear;
+}
+
+@keyframes scroll{
+    0%   {left: 500px;}
+    100% {left: -950px;}
 }
