+++ index.html
@@ -1,18 +1,15 @@
-
 <html>
-<head>
-
-<link rel="stylesheet" type="text/css" href="design.css">
-
-
-</head>
-
-<body>
-  <div id="points"  ></div>
+  <head>
+    <meta charset="utf-8">
+    <link rel="stylesheet" href="css/styles.css"/>
+    <link rel="stylesheet" href="node_modules/bootstrap/dist/css/bootstrap.min.css">
+  </head>
+  <body>
+<div id="points"  ></div>
    
   <button id="clicker" class="earth"></button>
   
-  	<button id="purchaseTree1" class="button1">Unlock after 3</button>
+    <button id="purchaseTree1" class="button1">Unlock after 3</button>
     
     <button id="purchaseCar1" class="button1">Electric Cars: Unlock after 50</button>
     
@@ -25,17 +22,17 @@
     <button id="purchaseCapture1" class="button1">Carbon Catcher: Unlock after 1000</button>
 
  <div id="newsfeed">
-	Scientists predict that life as we know it may cease to exist later this year!
+  Scientists predict that life as we know it may cease to exist later this year!
 </div>
     
     <div id="pollution" class="pollution">
-    <img style="height:auto; width:auto; max-height:400px; max-width:400px" src="1.png" class="image" />
-	</div>
+    <img style="height:auto; width:auto; max-height:400px; max-width:400px" src="img/1.png" class="image" />
+  </div>
     
     <div class="scrolling-limit"></div>
-    
-    <script src="Game.js"></script>
-    
-</body>
+    <script type="text/javascript" src="build/bundle.js" charset="utf-8"></script>
 
+    <script type="text/javascript" src="node_modules/jquery/dist/jquery.min.js"></script>
+    <script type="text/javascript" src="node_modules/bootstrap/dist/js/bootstrap.min.js"></script>
+  </body>
 </html>
